dziala migracja

+++ Lab1/Application/Services/CategoryServices.cs
@@ -1,33 +1,36 @@
-
+using Contacts.Application.Repositories;
 using Contacts.Application.UseCases;
 using Contacts.Application.UseCases.DTOs;
-using Contacts.Application.Repositories;
 
 namespace Contacts.Application.Services;
 
-public class CategoryServices(
-ICategoryRepository categoryRepository, IUnitOfWork unitOfWork
-) : ICategoryUseCases
+public class CategoryServices(ICategoryRepository categoryRepository, IUnitOfWork unitOfWork)
+    : ICategoryUseCases
 {
     public void AddCategory(AddCategoryDTO categoryDTO)
     {
         if (categoryRepository.GetCategoryByName(categoryDTO.Name) == null)
         {
-            categoryRepository.Add(categoryDTO); unitOfWork.Save();
+            categoryRepository.Add(categoryDTO);
+            unitOfWork.Save();
+        }
+        else
+        {
+            throw new ServiceException("Kategoria o tej nazwie już istnieje");
         }
     }
-else
-{
-}
-throw new ServiceException("Kategoria_o_tej_nazwie_już_istnieje");
+
     public void RemoveCategory(int categoryID)
     {
-        var category = categoryRepository.GetCategory(categoryID); if (category != null)
+        var category = categoryRepository.GetCategory(categoryID);
+        if (category != null)
         {
             categoryRepository.RemoveCategory(category);
             unitOfWork.Save();
         }
         else
         {
-            throw new ServiceException(
-            $"Kategoria_id:_{categoryID}_nie_istnieje");
+            throw new ServiceException($"Kategoria o id: {categoryID} nie istnieje");
+        }
+    }
+}
